pyproject.toml
[tool.poetry]
name = "energy-company"
version = "0.1.5"
description = ""
readme = "README.md"
packages = [{ include = "src" }]
authors = ["Pragash Rajarathna <pragash.r@petvisor.com>"]

[tool.poetry.dependencies]
fastapi = "^0.110.1"
httpx = "^0.27.0"
loguru = "^0.7.2"
python = "^3.12"
uvicorn = "^0.29.0"
databases = { extras = ["aiomysql"], version = "0.9.0" }
aiomysql = "0.2.0"
sqlalchemy = ">=2.0.7,<3.0"
cryptography = "^42.0.5"
python-dotenv = "^1.1.1"

[tool.poetry.group.tests]
optional = true
[tool.poetry.group.tests.dependencies]
pytest = "^8.1.1"
pytest-cov = "^5.0.0"
pytest-asyncio = "^0.23.6"

[tool.poetry.group.ci]
optional = true
[tool.poetry.group.ci.dependencies]
ruff = "^0.3.5"
black = "^24.3.0"

[tool.poetry.group.dev.dependencies]
pytest-asyncio = "^0.23.6"

[tool.pytest.ini_options]
asyncio_mode = "auto"
filterwarnings = [
    "ignore:.*Duplicate entry.*:Warning"
]

[[tool.poetry.source]]
name = "PyPI"
priority = "primary"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
lint.select = ["I00", "E", "F", "W"]

[tool.black]
line-length = 120
target-version = ['py39']
